+++ src/OpenTelemetry.Extensions.AWS/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* test: ensure aws xray propagator maintains context from previous propagator
+  ([#3145](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3145))
+
 ## 1.12.1
 
 Released 2025-Sep-03